requirements.txt
Django==1.8.5
djangorestframework==3.3.0
jsonfield==1.0.3
requests==2.8.1
wheel==0.24.0
